Update console file naming to be camelCase like
