+++ console-webapp/src/app/home/widgets/eppWidget.component.html
