+++ console-webapp/src/app/home/widgets/promotionsWidget.component.html
