+++ console-webapp/src/app/home/widgets/resourcesWidget.component.html
