+++ console-webapp/src/app/home/widgets/tldsWidget.component.ts
@@ -13,12 +13,11 @@
 // limitations under the License.
 
 import { Component } from '@angular/core';
-import { UserDataService } from 'src/app/shared/services/userData.service';
 
 @Component({
-  selector: '[app-resources-widget]',
-  templateUrl: './resources-widget.component.html',
+  selector: '[app-tlds-widget]',
+  templateUrl: './tldsWidget.component.html',
 })
-export class ResourcesWidgetComponent {
-  constructor(public userDataService: UserDataService) {}
+export class TldsWidgetComponent {
+  constructor() {}
 }
