+++ console-webapp/src/app/registrar/registrarSelector.component.html
