+++ console-webapp/src/app/registrar/registrarSelector.component.scss
